private-web: surface tailnet identity on server view; show alert_when_down state and unmonitored chip on child rows
